+++ cms-api/src/main/java/com/condation/cms/api/Constants.java
@@ -113,7 +113,7 @@ public abstract class Constants {
 	public static final int DEFAULT_PAGE_SIZE = 5;
 	
 	public static final String DEFAULT_CONTENT_TYPE = ContentTypes.HTML;
-	public static final List<String> DEFAULT_CONTENT_PIPELINE = List.of("markdown", "shortcode");
+	public static final List<String> DEFAULT_CONTENT_PIPELINE = List.of("markdown", "tags");
 	
 	public static final int DEFAULT_REDIRECT_STATUS = 301;