+++ cms-api/src/main/java/com/condation/cms/api/annotations/Tag.java
@@ -33,6 +33,6 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.METHOD})
-public @interface ShortCode {
+public @interface Tag {
 	String value ();
 }