+++ cms-content/src/main/java/com/condation/cms/content/tags/TagParser.java
@@ -1,4 +1,4 @@
-package com.condation.cms.content.shortcodes;
+package com.condation.cms.content.tags;
 
 /*-
  * #%L